avformat/vsf: add casting inside probe()

+++ libavformat/vsf.c
@@ -30,7 +30,7 @@ static int vsf_probe(const AVProbeData *p)
         return 0;
     if (p->buf_size < 32)
         return 0;
-    if (AV_RL32(p->buf + 28) <= 0)
+    if ((int)AV_RL32(p->buf + 28) <= 0)
         return 0;
 
     return AVPROBE_SCORE_MAX/3*2;
